+++ demo/app/components/demo-map/demo-map.component.ts
@@ -1,9 +1,9 @@
-import { Component, ContentChildren, Input, OnInit, QueryList, ViewChild, ViewChildren } from '@angular/core';
+import { Component, Input, ViewChild } from '@angular/core';
 import { ViewerConfiguration } from '../../../../src/services/viewer-configuration/viewer-configuration.service';
 import { AppSettingsService } from '../../services/app-settings-service/app-settings-service';
 import { MapLayerProviderOptions } from '../../../../src/models/map-layer-provider-options.enum';
-import { AcLayerComponent } from '../../../../src/components/ac-layer/ac-layer.component';
 import { TracksLayerComponent } from '../tracks-layer/tracks-layer.component';
+
 @Component({
   selector : 'demo-map',
   templateUrl : './demo-map.component.html',
@@ -12,10 +12,10 @@ import { TracksLayerComponent } from '../tracks-layer/tracks-layer.component';
 })
 export class DemoMapComponent {
   @Input() tracksRealData: boolean;
-
+  
   @ViewChild('layer') tracksLayer: TracksLayerComponent;
   arcGisMapServerProvider = MapLayerProviderOptions.ArcGisMapServer;
-
+  
   constructor(private viewerConf: ViewerConfiguration, private appSettingsService: AppSettingsService) {
     viewerConf.viewerOptions = {
       selectionIndicator : false,
@@ -28,15 +28,15 @@ export class DemoMapComponent {
       navigationHelpButton : false,
       navigationInstructionsInitiallyVisible : false,
     };
-
+    
     viewerConf.viewerModifier = (viewer) => {
       viewer.screenSpaceEventHandler.removeInputAction(Cesium.ScreenSpaceEventType.LEFT_DOUBLE_CLICK);
       viewer.bottomContainer.remove();
     };
-
+    
     this.appSettingsService.showTracksLayer = true;
-	}
-
+  }
+  
   removeAll() {
     this.tracksLayer.removeAll();
   }