+++ src/components/ac-map/ac-map.component.ts
@@ -10,10 +10,16 @@ import { DynamicPolylineDrawerService } from '../../services/drawers/dynamic-pol
 import { DynamicEllipseDrawerService } from '../../services/drawers/ellipse-drawer/dynamic-ellipse-drawer.service';
 import { PointDrawerService } from '../../services/drawers/point-drawer/point-drawer.service';
 import { ArcDrawerService } from '../../services/drawers/arc-drawer/arc-drawer.service';
+import { ViewersManagerService } from '../../services/viewers-service/viewers-manager.service';
 
 /**
  * This is a map implementation, creates the cesium map.
  * Every layer should be tag inside ac-map tag
+ *
+ * Accessing cesium viewer:
+ * 1. acMapComponent.getViewer()
+ * 2. Use ViewerManagerService.getCesiumViewer(mapId).
+ * 		mapId auto-generated string: 'default-map-id-[index]'
  *
  * @example
  * <ac-map>
@@ -34,6 +40,7 @@ export class AcMapComponent implements OnChanges, OnInit {
 	private static readonly DEFAULT_MINIMUM_ZOOM = 1.0;
 	private static readonly DEFAULT_MAXIMUM_ZOOM = Number.POSITIVE_INFINITY;
 	private static readonly DEFAULT_TILT_ENABLE = true;
+	private static defaultIdCounter = 1;
   
   /**
 	 * Disable default plonter context menu
@@ -61,6 +68,14 @@ export class AcMapComponent implements OnChanges, OnInit {
 	 */
 	@Input()
 	enableTilt: boolean = AcMapComponent.DEFAULT_TILT_ENABLE;
+ 
+	/**
+	 * Set the id name of the map
+	 * default: 'default-map-id-[index]'
+   * @type {string}
+   */
+	@Input()
+	id;
 
 	/**
 	 * flyTo options according to https://cesiumjs.org/Cesium/Build/Documentation/Camera.html?classFilter=cam#flyTo
@@ -70,10 +85,12 @@ export class AcMapComponent implements OnChanges, OnInit {
 
 	private mapContainer: HTMLElement;
 
-	constructor(private _cesiumService: CesiumService, private _elemRef: ElementRef, @Inject(DOCUMENT) private document: any) {
+	constructor(private _cesiumService: CesiumService,
+							private _elemRef: ElementRef,
+							@Inject(DOCUMENT) private document: any,
+							private viewersManager: ViewersManagerService) {
 		this.mapContainer = this.document.createElement('div');
 		this.mapContainer.className = 'map-container';
-		// this.mapContainer.style.height = '100%';
 		this._elemRef.nativeElement.appendChild(this.mapContainer);
 		this._cesiumService.init(this.mapContainer);
 	}
@@ -82,11 +99,23 @@ export class AcMapComponent implements OnChanges, OnInit {
 		this._cesiumService.setMinimumZoom(this.minimumZoom);
 		this._cesiumService.setMaximumZoom(this.maximumZoom);
 		this._cesiumService.setEnableTilt(this.enableTilt);
+		this.viewersManager.setViewer(this.id, this.getCesiumViewer());
 	}
 
 	ngOnChanges(changes: SimpleChanges): void {
 		if (changes['flyTo']) {
 			this._cesiumService.flyTo(changes['flyTo'].currentValue);
 		}
+	}
+  
+  /**
+   * @returns {any} map cesium viewer
+   */
+	getCesiumViewer() {
+		return this._cesiumService.getViewer();
+	}
+	
+	getId() {
+		return this.id;
 	}
 }